+++ .run/PowerauthFido2TestApplication.run.xml
@@ -2,9 +2,9 @@
   <configuration default="false" name="PowerauthFido2DemoApplication" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot" nameIsGenerated="true">
     <option name="ACTIVE_PROFILES" value="dev" />
     <option name="DEBUG_MODE" value="true" />
-    <module name="powerauth-fido2-demo" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="com.wultra.app.powerauth.fido2.demo.PowerauthFido2DemoApplication" />
-    <option name="VM_PARAMETERS" value="-Dserver.servlet.context-path=/powerauth-fido2-demo -Dserver.port=8083" />
+    <module name="powerauth-fido2-tests" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="com.wultra.security.powerauth.fido2.PowerauthFido2TestApplication" />
+    <option name="VM_PARAMETERS" value="-Dserver.servlet.context-path=/powerauth-fido2-test -Dserver.port=8083" />
     <option name="WORKING_DIRECTORY" value="file://$MODULE_WORKING_DIR$" />
     <method v="2">
       <option name="Make" enabled="true" />
