+++ docs-private/Developer-How-To-Start.md
@@ -19,11 +19,11 @@ Others (like URL, username, password) depend on your environment.
 liquibase --changelog-file=./docs/db/changelog/changesets/powerauth-test-server/db.changelog-module.xml --url=jdbc:postgresql://localhost:5432/powerauth --username=powerauth status
 ```
 
-## PowerAuth FIDO2 Demo
+## PowerAuth FIDO2 Tests
 
 ### Standalone Run
 
 - Enable maven profile `standalone`
 - Use IntelliJ Idea run configuration at `../.run/PowerAuthFido2DemoApplication.run.xml`
-- Open [http://localhost:8083/powerauth-fido2-demo/actuator/health](http://localhost:8083/powerauth-fido2-demo/actuator/health) and you should get `{"status":"UP"}`
+- Open [http://localhost:8083/powerauth-fido2-test/actuator/health](http://localhost:8083/powerauth-fido2-test/actuator/health) and you should get `{"status":"UP"}`
