--- powerauth-fido2-demo/src/main/resources/banner.txt
@@ -1,8 +1,0 @@
- ___                     _       _   _      ___ ___ ___   ___ ___   ___
- | _ \_____ __ _____ _ _ /_\ _  _| |_| |_   | __|_ _|   \ / _ \_  ) |   \ ___ _ __  ___
- |  _/ _ \ V  V / -_) '_/ _ \ || |  _| ' \  | _| | || |) | (_) / /  | |) / -_) '  \/ _ \
- |_| \___/\_/\_/\___|_|/_/ \_\_,_|\__|_||_| |_| |___|___/ \___/___| |___/\___|_|_|_\___/
-
-${AnsiColor.GREEN} :: ${banner.application.name} (${banner.application.version}) :: ${AnsiColor.GREEN}
-${AnsiColor.RED} :: Spring Boot${spring-boot.formatted-version} :: ${AnsiColor.RED}
-${AnsiColor.DEFAULT}
