+++ powerauth-fido2-tests/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 
 # Application Service Configuration
-powerauth.fido2.demo.service.applicationName=powerauth-fido2-demo
-powerauth.fido2.demo.service.applicationDisplayName=PowerAuth FIDO2 Demo
+powerauth.fido2.demo.service.applicationName=powerauth-fido2-tests
+powerauth.fido2.demo.service.applicationDisplayName=PowerAuth FIDO2 Test
 powerauth.fido2.demo.service.applicationEnvironment=
 
 banner.application.name=${powerauth.fido2.demo.service.applicationName}
