+++ powerauth-fido2-tests/src/main/resources/banner.txt
@@ -1,0 +1,8 @@
+  ___                     _       _   _      ___ ___ ___   ___ ___   _____       _
+ | _ \_____ __ _____ _ _ /_\ _  _| |_| |_   | __|_ _|   \ / _ \_  ) |_   _|__ __| |_
+ |  _/ _ \ V  V / -_) '_/ _ \ || |  _| ' \  | _| | || |) | (_) / /    | |/ -_|_-<  _|
+ |_| \___/\_/\_/\___|_|/_/ \_\_,_|\__|_||_| |_| |___|___/ \___/___|   |_|\___/__/\__|
+
+${AnsiColor.GREEN} :: ${banner.application.name} (${banner.application.version}) :: ${AnsiColor.GREEN}
+${AnsiColor.RED} :: Spring Boot${spring-boot.formatted-version} :: ${AnsiColor.RED}
+${AnsiColor.DEFAULT}
